search-analytics/mlapi/requirements.txt
annotated-types==0.6.0
async-timeout==4.0.3
blinker==1.7.0
click==8.1.7
Flask==3.0.2
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
numpy==1.26.4
packaging==23.2
psycopg2-binary==2.9.9
scipy==1.12.0
typing_extensions==4.10.0
Werkzeug==3.0.1
